qcacld-3.0: Validate session ID when going to suspend
When the session ID is invalid, it can cause an out-of-bounds read which
may or may not result in a panic, depending on the compiler and
compilation options used. Fix it by validating the session ID prior to
using it.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
(cherry picked from commit c90cd48685b971c1b6815c30b2c019a98e2b1c90)
Signed-off-by: fadlyas07 <fadlyardhians@gmail.com>